# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: KhelSetu API
  version: 1.0.0
  description: >
    # KhelSetu API Documentation


    ## Platform Overview


    KhelSetu is a multi-tenant sports tournament management platform with **two
    distinct portals**:


    ### 🌐 Public Portal (No Login Required)

    The public portal is for **viewers and fans** who want to watch live
    matches, browse tournaments, and view standings. These endpoints are
    **completely open** — no authentication needed.


    - **Live match scores** — Real-time scores for any match

    - **Tournament browsing** — View all public tournaments

    - **Standings & fixtures** — League tables and match schedules

    - **Team & player profiles** — Roster and player information

    - **Global search** — Search across tournaments, teams, and players


    ### 🔒 SaaS Dashboard (Login Required)

    The dashboard is for **organization members** (admins, scorers, coaches) who
    manage tournaments, teams, players, and live scoring. **All dashboard
    endpoints require JWT authentication** and **organization membership**.


    - **Organization management** — Create/manage organizations, invite members

    - **Tournament management** — Create tournaments, generate fixtures

    - **Team & player management** — Roster management, player profiles

    - **Live scoring** — Real-time event-driven scoring (Football, Cricket,
    Basketball)

    - **Standings** — Auto-calculated league tables

...

...

...


...


...

...
    refresh token (7d)
...
paths:
...
  /api/analytics/dashboard/stats:
    get:
      summary: Get dashboard statistics
      tags:
        - Analytics
      security:
        - bearerAuth: []
      parameters:
        - name: organization_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Dashboard stats
  /api/analytics/tournaments/{id}/stats:
    get:
      summary: Get tournament analytics
      tags:
        - Analytics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Tournament stats
  /api/analytics/matches/{id}/stats:
    get:
      summary: Get match analytics
      tags:
        - Analytics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Match stats
  /api/analytics/teams/{id}/stats:
    get:
      summary: Get team analytics
      tags:
        - Analytics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Team stats
  /api/analytics/players/{id}/stats:
    get:
      summary: Get player analytics
      tags:
        - Analytics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Player stats
  /api/analytics/reports/custom:
    get:
      summary: Generate custom report
      tags:
        - Analytics
      security:
        - bearerAuth: []
      parameters:
        - name: organization_id
          in: query
          required: true
          schema:
            type: string
        - name: start_date
          in: query
          schema:
            type: string
        - name: end_date
          in: query
          schema:
            type: string
        - name: tournament_id
          in: query
          schema:
            type: string
        - name: export_format
          in: query
          schema:
            type: string
            enum:
              - json
              - csv
      responses:
        "200":
          description: Custom report
  /api/analytics/leaderboards:
    get:
      summary: Get leaderboards
      tags:
        - Analytics
      security:
        - bearerAuth: []
      parameters:
        - name: organization_id
          in: query
          required: true
          schema:
            type: string
        - name: tournament_id
          in: query
          schema:
            type: string
        - name: metric
          in: query
          schema:
            type: string
            enum:
              - goals
              - assists
              - points
              - wins
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Leaderboard data
  /api/audit/logs:
    get:
      summary: Get audit logs
      tags:
        - Audit
      security:
        - bearerAuth: []
      parameters:
        - name: organization_id
          in: query
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: start_date
          in: query
          schema:
            type: string
        - name: end_date
          in: query
          schema:
            type: string
        - name: user_id
          in: query
          schema:
            type: string
        - name: action_type
          in: query
          schema:
            type: string
        - name: entity_type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Audit logs
  /api/audit/entity/{type}/{id}/history:
    get:
      summary: Get entity history
      tags:
        - Audit
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
            enum:
              - tournament
              - team
              - player
              - match
              - organization
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Entity history
  /api/audit/user/{userId}/activity:
    get:
      summary: Get user activity
      tags:
        - Audit
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: User activity
...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT access token. Get it from POST /api/auth/login. Pass as:
        Authorization: Bearer <token>"
...
security:
  - bearerAuth: []
...